refactor(tests): split test_spond.py by domain + add Chat tests
The single test_spond.py had grown past ~1100 lines and mixed seven
unrelated concerns. Split into eight files keyed by what they
exercise; shared constants, fixtures, and the
`_SpondBase.require_authentication` monkey-patch move to conftest.py
where pytest auto-loads them before any test module imports.

Also lands the TestChat suite (5 tests for the typed Chat/Message
surface added in the previous commit), placed in test_messaging.py
alongside the existing TestSendMessage tests.

New layout:
- conftest.py        — shared infrastructure
- test_auth.py       — login + decorator metadata
- test_compat.py     — DictCompatModel shim + update-payload guards
- test_events.py     — Event read/write, OO methods, Match subclass
- test_export.py     — deprecated xlsx export wrapper
- test_groups.py     — Group read + Member/Guardian navigation
- test_messaging.py  — send_message + Chat/Message
- test_posts.py      — get_posts queries + caching

62 tests pass; no behavioural changes.
